refactor: :truck: move "Views" to /views as "RootViews"
